Gave up on emacs
For some reason, while I try to do rename a symbol on neovim lsp it
works but for some reason in emacs it just says the server does not
support it even though that is a lie. and also for some reason trying to
set my own custom lsp keybinds did not work as the documentation only
tells you how to switch the prefix for which key mode. I guess for me
the pain of trying to set up emacs outweigh the costs of using it. (I am
not going to use a premade emacs distro like Doom)