Neovim: Change formatting configuration
When I was customizing the 4coder editor my previous configuration would
autoformat the files I was in due to it pointing to a .clang-format file
this would cause issues for some reason. So I wanted to only format
files which had a .clang-format file but I could not do that easily
without a plugin.

 I also did not want to use LSP (1 Month NO LSP Challenge but other
 tools such as linters/formatters/typecheckers are allowed) so I just
 used conform.nvim which is pretty good for what I am trying to do.