Emacs: Add finalized config
I do not use LSP, it is a pain to setup to my liking in emacs
and I also disabled lsp for neovim since I might as well not use
it on both If I don't use it on one.

I do have buffer completion. I'll probably tweak that later.